+++ src/C++/Message.h
@@ -202,9 +202,9 @@ public:
     return m_validStructure;
   }
 
-  int bodyLength( int beginStringField = FIELD::BeginString, 
-                  int bodyLengthField = FIELD::BodyLength, 
-                  int checkSumField = FIELD::CheckSum ) const
+  size_t bodyLength( int beginStringField = FIELD::BeginString, 
+                     int bodyLengthField = FIELD::BodyLength, 
+                     int checkSumField = FIELD::CheckSum ) const
   { return m_header.calculateLength(beginStringField, bodyLengthField, checkSumField)
            + calculateLength(beginStringField, bodyLengthField, checkSumField)
            + m_trailer.calculateLength(beginStringField, bodyLengthField, checkSumField);
